Refactor test structure and small testing related bugs
