+++ src/nimbo/tests/aws/test_core.py
@@ -2,9 +2,9 @@ import pytest
 from click.testing import CliRunner
 
 from nimbo import CONFIG
+from nimbo.core.cloud_provider.provider_impl.aws.aws_provider import AwsProvider
 from nimbo.core.config import RequiredCase
-from nimbo.core.utils import get_image_id
-from nimbo.tests.utils import isolated_filesystem
+from nimbo.tests.aws.utils import isolated_filesystem
 
 
 @pytest.fixture
@@ -19,20 +19,21 @@ def test_get_image_id(runner: CliRunner):
 
     CONFIG.image = reference_image
     CONFIG.region_name = reference_region
-    get_image_id()
+
+    AwsProvider._get_image_id()
 
     CONFIG.image = "awdghiuadgwui"
     with pytest.raises(ValueError):
-        get_image_id()
+        AwsProvider._get_image_id()
 
     CONFIG.image = reference_image
     CONFIG.region_name = "us-east-2"
-    get_image_id()
+    AwsProvider._get_image_id()
 
     with pytest.raises(ValueError):
         CONFIG.region_name = "us-et-2"
-        get_image_id()
+        AwsProvider._get_image_id()
 
     CONFIG.image = "ami-198571934781039"
     CONFIG.region_name = reference_region
-    get_image_id()
+    AwsProvider._get_image_id()
